backend-fastapi/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
python-multipart
pydantic
alembic
requests